- Server doxyfile and some documentation for database - Make the server code -Wall -W warning-free and add -Werror to makefile to keep it that way.
Former-commit-id: 4f7c5db1f80cfaeda7f5d57e393c6c77b533f6fe

+++ server/src/common/arch.c
@@ -688,7 +688,7 @@ unsigned long hasharch(const char *str, int tablesize)
 	{
         hash ^= (unsigned long) *p << rot;
         rot += 2;
-        if (rot >= (sizeof(long) - sizeof(char)) * 8)
+        if (rot >= ((int) sizeof(long) - (int) sizeof(char)) * 8)
             rot = 0;
     }
