+++ server/src/common/item.c
@@ -241,7 +241,7 @@ char *describe_protections(object *op, int newline)
 char *query_weight(object *op)
 {
 	static char buf[10];
-	int i = op->nrof ? op->nrof * op->weight : op->weight + op->carrying;
+	int i = op->nrof ? (int) op->nrof * op->weight : op->weight + op->carrying;
 
 	if (op->weight < 0)
 		return "      ";
