+++ server/src/server/c_object.c
@@ -571,7 +571,7 @@ void put_object_in_sack(object *op, object *sack, object *tmp, long nrof)
     }
 	*/
 
-    if (!sack_can_hold(op, sack, tmp, (nrof ? nrof : tmp->nrof)))
+    if (!sack_can_hold(op, sack, tmp, (nrof ? nrof : (int) tmp->nrof)))
       	return;
 
     if (QUERY_FLAG(tmp, FLAG_APPLIED))
@@ -1634,7 +1634,7 @@ void examine(object *op, object *tmp)
 
     if (tmp->weight)
 	{
-		sprintf(buf, tmp->nrof > 1 ? "They weigh %3.3f kg." : "It weighs %3.3f kg.", (float)(tmp->nrof ? tmp->weight * tmp->nrof : tmp->weight) / 1000.0f);
+		sprintf(buf, tmp->nrof > 1 ? "They weigh %3.3f kg." : "It weighs %3.3f kg.", (float)(tmp->nrof ? tmp->weight * (int) tmp->nrof : tmp->weight) / 1000.0f);
 		new_draw_info(NDI_UNIQUE, 0, op, buf);
     }
